Put argparse specific arguments inside of bin
